phase2a/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phase2a"
version = "0.1.0"
description = "Automated satellite tracing for golf course feature extraction"
readme = "../README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["golf", "satellite", "segmentation", "svg", "gis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "scikit-image>=0.21.0",
    "shapely>=2.0.0",
    "geopandas>=0.14.0",
    "svgwrite>=1.4.3",
    "cairosvg>=2.7.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
sam = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "segment-anything @ git+https://github.com/facebookresearch/segment-anything.git",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = ["phase2a[sam,dev]"]

[project.scripts]
phase2a = "phase2a.cli:main"

[project.urls]
Homepage = "https://github.com/menkelabs/course-builder"
Repository = "https://github.com/menkelabs/course-builder"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
